backend/infrastructure/requirements.txt
# AWS CDK Core Library
aws-cdk-lib==2.120.0

# Constructs base library
constructs>=10.0.0,<11.0.0

# Python type checking (optional but recommended)
mypy==1.7.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
